fixed mysql's inappropriate handling of YEAR

+++ src/com/cooksys/ssi/models/JobIdentifier.scala
@@ -1,5 +1,5 @@
 package com.cooksys.ssi.models
 
 case class JobIdentifier(prefix: JobPrefix,
-                         year: Int,
+                         year: Short,
                          label: String)
